requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
openai>=1.0.0
feedparser==6.0.10
html2text>=2025.4.15
psycopg2-binary>=2.9.9
